+++ src/app/admin/login/page.tsx
@@ -1,0 +1,68 @@
+"use client";
+
+import Link from "next/link";
+import React, { useState } from "react";
+
+export default function LoginPage() {
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+    // TODO: connect to API or auth system
+    console.log("Login attempt:", { email, password });
+  };
+
+  return (
+    <div
+      className="
+    flex items-center justify-center min-h-screen bg-gray-100"
+    >
+      <div className="w-full max-w-sm bg-white p-6 rounded-2xl shadow-md">
+        <h1 className="text-2xl font-bold text-center mb-6">Admin Login</h1>
+
+        <form onSubmit={handleSubmit} className="space-y-4">
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Email
+            </label>
+            <input
+              type="email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              required
+              className="mt-1 block w-full px-3 py-2 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-indigo-500"
+              placeholder="you@example.com"
+            />
+          </div>
+
+          <div>
+            <label className="block text-sm font-medium text-gray-700">
+              Password
+            </label>
+            <input
+              type="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              required
+              className="mt-1 block w-full px-3 py-2 border rounded-lg shadow-sm focus:outline-none focus:ring-2 focus:ring-indigo-500"
+              placeholder="••••••••"
+            />
+          </div>
+
+          <Link
+            href="/zz
+          "
+          >
+            <button
+              type="submit"
+              className="w-full bg-indigo-600 text-white py-2 rounded-lg hover:bg-indigo-700 transition"
+            >
+              Login
+            </button>
+          </Link>
+        </form>
+      </div>
+    </div>
+  );
+}
