--- src/app/admin/mail-management/page.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const page = () => {
-  return <div>page</div>;
-};
-
-export default page;
