+++ src/app/admin/page.tsx
@@ -1,19 +1,12 @@
-import Sidebar from "./components/Sidebar";
-import Dashboard from "./components/Dashboard";
-import BlogManagement from "./components/blogcomp/BlogManagement";
+import Dashboard from "./Dashboard/page";
 
 export default function Home() {
   return (
     <div className="flex bg-gray-100">
-      {/* Sidebar */}
-      <div className="w-64">
-        <Sidebar />
-      </div>
-
       {/* Main Content */}
       <div className="flex-1 h-screen overflow-y-auto">
-       {/*  <Dashboard /> */}
-         <BlogManagement/> 
+        <Dashboard />
+        {/* <BlogManagement /> */}
       </div>
     </div>
   );
