--- src/lib/models/Blog.ts
@@ -1,12 +1,0 @@
-import mongoose, { Schema, models } from "mongoose";
-
-const BlogSchema = new Schema(
-  {
-    title: { type: String, required: true },
-    content: { type: String, required: true },
-  },
-  { timestamps: true }
-);
-
-const Blog = models.Blog || mongoose.model("Blog", BlogSchema);
-export default Blog;
